RPCCommand: update macro + function signature changes
After a lot of testing and playing in macro land, it emerged to be that the most optimal way to do things with the framework's design in mind was to simply macro everything. This does unfortunately remove the ability to manage commands at runtime, but it is worth the rigid constraint resulting from the decision.