CrossByte: Main Loop + network changes
One of the things I wanted to do with CrossByte was to make the Main Loop entirely modular. Up until now, we havent had much of a reason to implement it, but here it is.

We kind of radically changed the way we handle sockets **(again!)** here, and even provide separate implementations based on the main loop type.